Get rid of global data -- all data is now per-processor
The MSR Bitmap and EPML4/EPDPTEs were allocated globally, with each LP's
VMCS pointing to the same physical address. Make things simpler and more
efficient (especially if we take advantage of NUMA) by giving each
LP/VMCS its own MSR Bitmap and EPT structures. The intent here is not to
de-synchronize this data, as we have no exits which would cause either
of these structures to change and require synchronization.